Move dashboard overview and apps to new navigation
