+++ app/src/main/java/com/nasdroid/MainActivity.kt
@@ -25,6 +25,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.nasdroid.apps.ui.appsGraph
@@ -51,149 +52,39 @@ class MainActivity : ComponentActivity() {
         setContent {
             val windowSizeClass = calculateWindowSizeClass(activity = this)
             NasDroidTheme {
-                MainScreen(windowSizeClass)
+                AppContent(windowSizeClass)
             }
         }
     }
 }
 
-/**
- * The main content of the app. This Composable displays navigation-related elements, and contains a
- * NavHost for features to display their own screens.
- */
 @Composable
-fun MainScreen(
-    windowSizeClass: WindowSizeClass
-) {
-    val destinations = remember {
-        TopLevelDestination.entries
-    }
-
+fun AppContent(windowSizeClass: WindowSizeClass) {
     val navController = rememberNavController()
-    val currentBackstackEntry by navController.currentBackStackEntryAsState()
-    val selectedDestination by remember {
-        derivedStateOf {
-            val currentRoute = currentBackstackEntry?.destination?.parent?.route
-            destinations.firstOrNull { it.getRoute() == currentRoute }
-        }
-    }
-
-    var isPowerOptionsVisible by rememberSaveable { mutableStateOf(false) }
-
-    NavigationDrawerLayout(
-        drawerHeaderContent = {
-            DrawerServerSelector(
-                onLogout = {
-                    navController.navigate("auth") {
-                        selectedDestination?.getRoute()?.let { selectedDestinationRoute ->
-                            popUpTo(selectedDestinationRoute) {
-                                inclusive = true
-                            }
-                        }
-                    }
-                },
-                modifier = Modifier.padding(horizontal = MaterialThemeExt.paddings.large),
-                additionalControls = {
-                    OutlinedIconButton(onClick = { isPowerOptionsVisible = true }) {
-                        Icon(Icons.Default.SettingsPower, "Power")
-                    }
-                }
-            )
-        },
-        onNavigationItemClick = { navController.navigate(it.route) }
-    ) {
-        MainNavHost(
-            navController = navController,
-            destinations = destinations,
-            modifier = Modifier
-                .fillMaxSize(),
-            contentPadding = PaddingValues(horizontal = 16.dp, vertical = 12.dp),
-            windowSizeClass = windowSizeClass,
-        )
-    }
-    if (isPowerOptionsVisible) {
-        PowerOptionsDialog(
-            onDismiss = { isPowerOptionsVisible = false }
-        )
-    }
-}
-
-/**
- * Shows a [NavHost] with routes for all given [destinations].
- */
-@Composable
-fun MainNavHost(
-    navController: NavHostController,
-    windowSizeClass: WindowSizeClass,
-    destinations: List<TopLevelDestination>,
-    modifier: Modifier = Modifier,
-    contentPadding: PaddingValues = PaddingValues(),
-) {
-    NavHost(
-        navController = navController,
-        startDestination = "auth",
-        modifier = modifier
-    ) {
+    NavHost(navController, "auth") {
         authNavigation(
             navController = navController,
             route = "auth",
+            windowSizeClass = windowSizeClass,
             onLoginSuccess = {
-                navController.navigate(TopLevelDestination.Dashboard.getRoute()) {
+                navController.navigate("content") {
                     popUpTo("auth") {
                         inclusive = true
                     }
                 }
-            },
-            contentPadding = contentPadding,
-            windowSizeClass = windowSizeClass,
-        )
-
-        destinations.forEach { destination ->
-            when (destination) {
-                TopLevelDestination.Dashboard -> dashboardGraph(
-                    route = destination.getRoute(),
-                    contentPadding = contentPadding
-                )
-                TopLevelDestination.Storage -> storageGraph(
-                    windowSizeClass = windowSizeClass,
-                    navController = navController,
-                    route = destination.getRoute(),
-                    contentPadding = contentPadding
-                )
-                TopLevelDestination.Datasets -> {}
-                TopLevelDestination.Shares -> {}
-                TopLevelDestination.DataProtection -> {}
-                TopLevelDestination.Network -> {}
-                TopLevelDestination.Credentials -> {}
-                TopLevelDestination.Virtualization -> {}
-                TopLevelDestination.Apps -> appsGraph(
-                    windowSizeClass = windowSizeClass,
-                    navController = navController,
-                    route = destination.getRoute(),
-                    contentPadding = contentPadding
-                )
-                TopLevelDestination.Reporting -> reportingGraph(
-                    route = destination.getRoute(),
-                    contentPadding = contentPadding
-                )
-                TopLevelDestination.SystemSettings -> {}
             }
+        )
+        composable("content") {
+            MainScreen(
+                onLogOut = {
+                    navController.navigate("auth") {
+                        popUpTo("auth") {
+                            inclusive = true
+                        }
+                    }
+                },
+                windowSizeClass = windowSizeClass
+            )
         }
-    }
-}
-
-internal fun TopLevelDestination.getRoute(): String {
-    return when (this) {
-        TopLevelDestination.Dashboard -> "dashboard"
-        TopLevelDestination.Storage -> "storage"
-        TopLevelDestination.Datasets -> "datasets"
-        TopLevelDestination.Shares -> "shares"
-        TopLevelDestination.DataProtection -> "data_protection"
-        TopLevelDestination.Network -> "network"
-        TopLevelDestination.Credentials -> "credentials"
-        TopLevelDestination.Virtualization -> "virtualization"
-        TopLevelDestination.Apps -> "apps"
-        TopLevelDestination.Reporting -> "reporting"
-        TopLevelDestination.SystemSettings -> "system_settings"
     }
 }
