+++ features/dashboard/ui/src/main/kotlin/com/nasdroid/dashboard/ui/overview/DashboardOverviewScreen.kt
@@ -3,6 +3,7 @@ package com.nasdroid.dashboard.ui.overview
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.staggeredgrid.LazyVerticalStaggeredGrid
 import androidx.compose.foundation.lazy.staggeredgrid.StaggeredGridCells
 import androidx.compose.foundation.lazy.staggeredgrid.itemsIndexed
@@ -28,6 +29,7 @@ import com.nasdroid.dashboard.ui.overview.cpu.CpuOverview
 import com.nasdroid.dashboard.ui.overview.memory.MemoryOverview
 import com.nasdroid.dashboard.ui.overview.network.NetworkOverview
 import com.nasdroid.dashboard.ui.overview.system.SystemInformationOverview
+import com.nasdroid.navigation.NavigationSuiteScaffold
 import org.koin.androidx.compose.koinViewModel
 
 /**
@@ -45,32 +47,37 @@ fun DashboardOverviewScreen(
     val isEditing by remember(editingItems) {
         derivedStateOf { editingItems != null }
     }
-    items?.getOrNull()?.let { dashboardItems ->
-        ProvideMenuItems(
-            if (isEditing) {
-                MenuItem(
-                    label = "Stop Editing",
-                    imageVector = Icons.Default.EditOff,
-                    onClick = viewModel::stopEditing,
-                    isImportant = true,
-                )
-            } else {
-                MenuItem(
-                    label = "Edit",
-                    imageVector = Icons.Default.Edit,
-                    onClick = viewModel::startEditing,
-                    isImportant = true,
-                )
-            }
-        )
-        DashboardOverviewList(
-            items = editingItems ?: dashboardItems,
-            isEditing = isEditing,
-            contentPadding = contentPadding,
-            onMoveItem = viewModel::moveDashboardEntry,
-            onStartEditing = viewModel::startEditing,
-            modifier = modifier
-        )
+    NavigationSuiteScaffold(
+        title = { Text("Dashboard") },
+        modifier = modifier
+    ) {
+        items?.getOrNull()?.let { dashboardItems ->
+            ProvideMenuItems(
+                if (isEditing) {
+                    MenuItem(
+                        label = "Stop Editing",
+                        imageVector = Icons.Default.EditOff,
+                        onClick = viewModel::stopEditing,
+                        isImportant = true,
+                    )
+                } else {
+                    MenuItem(
+                        label = "Edit",
+                        imageVector = Icons.Default.Edit,
+                        onClick = viewModel::startEditing,
+                        isImportant = true,
+                    )
+                }
+            )
+            DashboardOverviewList(
+                items = editingItems ?: dashboardItems,
+                isEditing = isEditing,
+                contentPadding = contentPadding,
+                onMoveItem = viewModel::moveDashboardEntry,
+                onStartEditing = viewModel::startEditing,
+                modifier = Modifier.padding(it)
+            )
+        }
     }
 }
